requirements.txt
Django
djangorestframework
openai
firebase_admin
requests
